- Extract detailed status from OPC return codes.

+++ src/main/java/com/hurence/opc/da/OpcDaOperations.java
@@ -273,7 +273,7 @@ public class OpcDaOperations extends AbstractOpcOperations<OpcDaConnectionProfil
                                 ret.setProperties(new HashSet<>(tagProps.values()));
                                 //set common properties
                                 if (tagProps.containsKey(OpcDaItemProperties.MANDATORY_DATA_TYPE)) {
-                                    OpcTagProperty<Integer> tmp = tagProps.get(OpcDaItemProperties.MANDATORY_DATA_TYPE);
+                                    OpcTagProperty<Short> tmp = tagProps.get(OpcDaItemProperties.MANDATORY_DATA_TYPE);
                                     ret.setType(JIVariantMarshaller.findJavaClass(tmp != null && tmp.getValue() != null ? tmp.getValue() : JIVariant.VT_EMPTY));
                                 }
